+++ packages/main/src/Button.ts
@@ -426,7 +426,10 @@ class Button extends UI5Element implements IButton, IFormElement {
 		if (this._cancelAction) {
 			e.preventDefault();
 		}
-		markEvent(e, "button");
+
+		if (isSpace(e)) {
+			markEvent(e, "button");
+		}
 
 		if (isSpace(e) || isEnter(e)) {
 			if (this.active) {
@@ -527,6 +530,10 @@ class Button extends UI5Element implements IButton, IFormElement {
 
 	get ariaLabelText() {
 		return getEffectiveAriaLabelText(this);
+	}
+
+	get ariaDescribedbyText() {
+		return this.hasButtonType ? "ui5-button-hiddenText-type" : undefined;
 	}
 
 	get _isSubmit() {